pyproject.toml
[tool.poetry]
name = "bobiversestmap"
version = "0.1.0"
description = ""
authors = ["Matthieu Berthomé <matthieu@mmea.fr>"]

[tool.poetry.dependencies]
python = "^3.8.5"
flask-nav = "*"
Flask = "*"
Flask-Bootstrap = "*"
Flask-Script = "*"
Flask-SQLAlchemy = "*"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]
nltk = "*"
python-dateutil = "^2.8.1"
colorcet = "*"
flask-nav = "*"
"dogpile.cache" = "*"
astropy = "*"
sortedcontainers = "*"
Flask = "*"
Frozen-Flask = "*"
Flask-Bootstrap = "*"
Flask-Script = "*"
Flask-SQLAlchemy = "*"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
